+++ CHANGELOG.md
@@ -4,6 +4,10 @@
 
 ## [Unreleased]
 
+### Changed
+
+- Fallback to primitive string type for malformed JSON and JSON5 inputs
+
 ## [5.0.1] - 2026-04-19
 
 ### Changed